Add some analytics points for refactor
