+++ calva/refactor/index.ts
@@ -1,6 +1,6 @@
 import * as vscode from 'vscode';
 import * as impl from './impl';
-import CalvaReferenceProvider from "../providers/reference";
+import CalvaReferenceProvider from "./providers/reference";
 import * as state from '../state';
 
 function activate(context: vscode.ExtensionContext) {
